+++ metod1/metod1/tetrahedron_atd.h
@@ -4,7 +4,7 @@
 namespace simple_shapes {
 	// шар
 	struct tetrahedron {
-		float l; // длина стороны
+		int l; // длина стороны
 		float d; //плотность(density)
 	};
 } // end simple_shapes namespace
